Begin moving repository operations into separate package
