+++ internal/config/config.go
@@ -4,21 +4,12 @@ import (
 	"os"
 
 	"github.com/pelletier/go-toml/v2"
+	"go.arsenm.dev/lure/internal/types"
 )
 
-type Config struct {
-	RootCmd string `toml:"rootCmd"`
-	Repos   []Repo `toml:"repo"`
-}
-
-type Repo struct {
-	Name string `toml:"name"`
-	URL  string `toml:"url"`
-}
-
-var defaultConfig = Config{
+var defaultConfig = types.Config{
 	RootCmd: "sudo",
-	Repos: []Repo{
+	Repos: []types.Repo{
 		{
 			Name: "default",
 			URL:  "https://github.com/Arsen6331/lure-repo.git",
@@ -26,7 +17,7 @@ var defaultConfig = Config{
 	},
 }
 
-func Decode(cfg *Config) error {
+func Decode(cfg *types.Config) error {
 	cfgFl, err := os.Open(ConfigPath)
 	if err != nil {
 		return err
