+++ internal/config/version.go
@@ -1,0 +1,6 @@
+package config
+
+import _ "embed"
+
+//go:embed version.txt
+var Version string
