+++ internal/types/repo.go
@@ -1,0 +1,7 @@
+package types
+
+type RepoConfig struct {
+	Repo struct {
+		MinVersion string `toml:"minVersion"`
+	}
+}
