+++ repo.go
@@ -19,19 +19,16 @@
 package main
 
 import (
-	"context"
-	"net/url"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/go-git/go-git/v5"
 	"github.com/pelletier/go-toml/v2"
 	"github.com/urfave/cli/v2"
 	"go.arsenm.dev/logger/log"
-	"go.arsenm.dev/lure/download"
 	"go.arsenm.dev/lure/internal/config"
+	"go.arsenm.dev/lure/internal/repos"
+	"go.arsenm.dev/lure/internal/types"
 	"golang.org/x/exp/slices"
 )
 
@@ -41,12 +38,6 @@ type PkgNotFoundError struct {
 
 func (p PkgNotFoundError) Error() string {
 	return "package '" + p.pkgName + "' could not be found in any repository"
-}
-
-type RepoConfig struct {
-	Repo struct {
-		MinVersion string `toml:"minVersion"`
-	}
 }
 
 func addrepoCmd(c *cli.Context) error {
@@ -59,7 +50,7 @@ func addrepoCmd(c *cli.Context) error {
 		}
 	}
 
-	cfg.Repos = append(cfg.Repos, config.Repo{
+	cfg.Repos = append(cfg.Repos, types.Repo{
 		Name: name,
 		URL:  repoURL,
 	})
@@ -113,7 +104,7 @@ func removerepoCmd(c *cli.Context) error {
 }
 
 func refreshCmd(c *cli.Context) error {
-	err := pullRepos(c.Context)
+	err := repos.Pull(c.Context, cfg.Repos)
 	if err != nil {
 		log.Fatal("Error pulling repos").Err(err).Send()
 	}
@@ -196,98 +187,4 @@ func findPkgs(pkgs []string) (scripts, notFound []string) {
 		}
 	}
 	return
-}
-
-func pullRepos(ctx context.Context) error {
-	for _, repo := range cfg.Repos {
-		repoURL, err := url.Parse(repo.URL)
-		if err != nil {
-			return err
-		}
-
-		log.Info("Pulling repository").Str("name", repo.Name).Send()
-		repoDir := filepath.Join(config.RepoDir, repo.Name)
-
-		gitDir := filepath.Join(repoDir, ".git")
-		if fi, err := os.Stat(gitDir); err == nil && fi.IsDir() {
-			r, err := git.PlainOpen(repoDir)
-			if err != nil {
-				return err
-			}
-
-			w, err := r.Worktree()
-			if err != nil {
-				return err
-			}
-
-			err = w.PullContext(ctx, &git.PullOptions{Progress: os.Stderr})
-			if err == git.NoErrAlreadyUpToDate {
-				log.Info("Repository up to date").Str("name", repo.Name).Send()
-				continue
-			} else if err != nil {
-				return err
-			}
-
-			fl, err := w.Filesystem.Open("lure-repo.toml")
-			if err != nil {
-				log.Warn("Git repository does not appear to be a valid LURE repo").Str("repo", repo.Name).Send()
-				continue
-			}
-
-			var repoCfg RepoConfig
-			err = toml.NewDecoder(fl).Decode(&repoCfg)
-			if err != nil {
-				return err
-			}
-			fl.Close()
-
-			currentVer, _, _ := strings.Cut(version, "-")
-			if vercmp(currentVer, repoCfg.Repo.MinVersion) == -1 {
-				log.Warn("LURE repo's minumum LURE version is greater than the current version. Try updating LURE if something doesn't work.").Str("repo", repo.Name).Send()
-			}
-
-			continue
-		}
-
-		err = os.RemoveAll(repoDir)
-		if err != nil {
-			return err
-		}
-
-		err = os.MkdirAll(repoDir, 0o755)
-		if err != nil {
-			return err
-		}
-
-		if !strings.HasPrefix(repoURL.Scheme, "git+") {
-			repoURL.Scheme = "git+" + repoURL.Scheme
-		}
-
-		err = download.Get(ctx, download.GetOptions{
-			SourceURL:   repoURL.String(),
-			Destination: repoDir,
-		})
-		if err != nil {
-			return err
-		}
-
-		fl, err := os.Open(filepath.Join(repoDir, "lure-repo.toml"))
-		if err != nil {
-			log.Warn("Git repository does not appear to be a valid LURE repo").Str("repo", repo.Name).Send()
-		}
-
-		var repoCfg RepoConfig
-		err = toml.NewDecoder(fl).Decode(&repoCfg)
-		if err != nil {
-			return err
-		}
-		fl.Close()
-
-		currentVer, _, _ := strings.Cut(version, "-")
-		if vercmp(currentVer, repoCfg.Repo.MinVersion) == -1 {
-			log.Warn("LURE repo's minumum LURE version is greater than the current version. Try updating LURE if something doesn't work.").Str("repo", repo.Name).Send()
-		}
-	}
-
-	return nil
 }
